refactor(air): split air.rs into dedicated modules
Split the monolithic `air.rs` (~500 lines) into focused modules
as proposed in #1434:

- `air.rs`: `BaseAir` and `Air` traits only
- `builder.rs`: `AirBuilder` and extension traits
  (`PeriodicAirBuilder`, `AirBuilderWithContext`, `ExtensionBuilder`,
  `PermutationAirBuilder`)
- `filtered.rs`: `FilteredAirBuilder` and its trait impls
- `window.rs`: `WindowAccess` trait and `RowWindow`

`utils.rs` and `virtual_column.rs` are unchanged.

All items are re-exported from `lib.rs`, so downstream imports
remain stable.

Closes #1434
