+++ air/src/lib.rs
@@ -5,14 +5,20 @@
 extern crate alloc;
 
 mod air;
+mod builder;
 mod check_constraints;
+mod filtered;
 mod named;
 pub mod symbolic;
 pub mod utils;
 mod virtual_column;
+mod window;
 
 pub use air::*;
+pub use builder::*;
 pub use check_constraints::*;
+pub use filtered::*;
 pub use named::*;
 pub use symbolic::*;
 pub use virtual_column::*;
+pub use window::*;
